ajout de l'authentification pour acceder à la page admin et l'api

+++ server/routes/messages.mjs
@@ -13,6 +13,9 @@ router.get("/", async (req, res) => {
 router.post("/", async (req, res) => {
   let collection = await db.collection("messages");
   let newDocument = req.body;
+  const userId = req.headers["user-id"];
+
+  if (userId !== "661151b364d136a4fd354ed1") return res.status(403);
   newDocument.date = new Date();
   let result = await collection.insertOne(newDocument);
   res.send(result).status(204);
@@ -22,6 +25,9 @@ router.delete("/:id", async (req, res) => {
   console.log("delete");
   const query = { _id: ObjectId(req.params.id) };
 
+  const userId = req.headers["user-id"];
+
+  if (userId !== "661151b364d136a4fd354ed1") return res.status(403);
   const collection = db.collection("messages");
   console.log(collection);
   let result = await collection.deleteOne(query);
@@ -32,6 +38,10 @@ router.delete("/:id", async (req, res) => {
 router.post("/archive", async (req, res) => {
   let collection = await db.collection("messagesArchived");
   let newDocument = req.body;
+
+  const userId = req.headers["user-id"];
+
+  if (userId !== "661151b364d136a4fd354ed1") return res.status(403);
   newDocument.date = new Date();
   let result = await collection.insertOne(newDocument);
   res.send(result).status(204);
